+++ addon/components/paper-sidenav-toggle.js
@@ -3,18 +3,14 @@ const { Component, inject } = Ember;
 
 export default Component.extend({
 
+  tagName: '',
+
   name: 'default',
 
   paperSidenav: inject.service(),
 
-  click() {
+  toggle() {
     this.get('paperSidenav').toggle(this.get('name'));
-  },
-
-  actions: {
-    toggleMenu() {
-      this.get('paperSidenav').toggle(this.get('name'));
-    }
   }
 
 });